Create utils/batch.py
